Organize tutorial images in separate folder

+++ Pencil_core/src/graphics/textures/texture.h
@@ -10,8 +10,8 @@ namespace Pencil {
 
 	class Texture {
 	private:
-		static short m_TexIndex;
-		short m_TextureSlot;
+		static GLushort m_TexIndex;
+        GLushort m_TextureSlot;
 		GLuint m_TextureId;
 		GLsizei m_Width, m_Height, m_NmbChannels;
 		std::string m_FileName;
@@ -23,7 +23,7 @@ namespace Pencil {
 		void unbind();
 
 		inline const GLuint getID() const { return m_TextureId;  }
-		inline const GLuint getTextureSlot() const { return (m_TextureSlot - GL_TEXTURE0); }
+		inline const GLushort getTextureSlot() const { return (m_TextureSlot - GL_TEXTURE0); }
 		inline const GLsizei getWidth() const { return m_Width;  }
 		inline const GLsizei getHeight() const { return m_Height;  }
